+++ weblate/trans/change_display.py
@@ -283,6 +283,7 @@ class RenderUserActions(BaseDetailsRenderStrategy):
         ActionEvents.ADD_USER,
         ActionEvents.INVITE_USER,
         ActionEvents.REMOVE_USER,
+        ActionEvents.USER_ACCESS_CHANGE,
         ActionEvents.USER_REVERT,
     }
     details_required = True